fix: cell constraints now follow each labels width
