+++ MacroChallengeTeam2/View/KeyboardCollectionViewCell.swift
@@ -9,28 +9,56 @@ import UIKit
 
 class KeyboardCollectionViewCell: UICollectionViewCell {
     static let identifier = "KeyboardCollectionViewCell"
-    
-    private lazy var snippetView = SnippetView()
+    var snippet: String?
+    var btnLabel: String? {
+        didSet {
+            label.text = btnLabel
+        }
+    }
+
+    private lazy var backBG: UIView = {
+        let view = UIView()
+        view.translatesAutoresizingMaskIntoConstraints = false
+        view.layer.cornerRadius = 5
+        view.isUserInteractionEnabled = false
+        view.backgroundColor = .white
+        return view
+    }()
+
+    private lazy var label: UILabel = {
+        let label = UILabel()
+        label.textColor = .kobarBlack
+        label.font = .regular17
+        label.textAlignment = .center
+        return label
+    }()
     
     
     override init(frame: CGRect) {
         super.init(frame: frame)
-        contentView.addSubview(snippetView)
-    }
-    
-    private func setupAutoLayout() {
-        snippetView.snp.makeConstraints { make in
-            make.height.equalToSuperview()
-            make.width.equalToSuperview()
-            make.center.equalToSuperview()
-        }
-    }
-    
-    func setSnippetView(title: String, snippet: String) {
-        snippetView.setSnippet(title: title, snippet: snippet)
+        contentView.addSubview(backBG)
+        contentView.addSubview(label)
+        self.clipsToBounds = true
         setupAutoLayout()
     }
     
+    private func setupAutoLayout() {
+        backBG.snp.makeConstraints { make in
+            make.height.equalTo(36)
+            make.width.equalTo(label.snp.width).offset(20)
+            make.center.equalToSuperview()
+        }
+        label.snp.makeConstraints { make in
+            make.height.equalTo(label.snp.height)
+            make.width.equalTo(label.snp.width)
+            make.center.equalToSuperview()
+        }
+        contentView.snp.makeConstraints { make in
+            make.width.equalTo(backBG.snp.width)
+            make.height.equalTo(backBG.snp.height)
+        }
+    }
+
     required init?(coder: NSCoder) {
         fatalError("error cok")
     }
